drivers: irqsteer: Make IRQSTEER_GetMasterNextInterrupt work on any architecture
Currently, IRQSTEER_GetMasterNextInterrupt works only on ARM-based
SoCs due to the fact that it uses __RBIT and __CLZ which are
ARM-specific. As such, make the function more generic so that it
can be used on any architecture. In the case of Cortex-M, this function
will use __RBIT and __CLZ instead of the generic approach as
it's more efficient to use these special instructions instead of
the generic for-loop.

Signed-off-by: Laurentiu Mihalcea <laurentiu.mihalcea@nxp.com>